requirements.txt
torch
numpy
pillow
matplotlib
imageio
pyyaml
